requirements.txt
flask==3.0.3
pyyaml==6.0.2
qdrant-client==1.9.1
requests==2.32.3
pytest==8.3.2
